Fix loss and accuracy logging during training
During training, loss and metrics (accuracy) are not displayed or
stored at the end of each epoch. As a result, training or validation
loss cannot be retrieved after training. Furthermore, callbacks such
as CSVLogger and EarlyStopping are not functional.

Modify 'on_epoch_end' in 'History' callback, and in 'fit' method to
enable logging of loss and metrics.

Modify CSVLogger such that output file header conforms to history
logging format.

Resolves: #75, #83
See also: #60